feat: refactor and protect routes
